create overview wrapper component
